+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED] - XXXX.XX.XX
+
+### Added
+
+- utils.py: add convert_rhcbci_to_rhntw for converting vectors from right-handed celestial-body-centered-inertial to right-handed NTW frame
+
 ## [v0.13.0] - 2025.10.17
 
 ### Added